% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISQG.R
\docType{data}
\name{ToyMap}
\alias{ToyMap}
\title{Toy Example of a Map for \emph{in silico} Quantitative Genetics}
\format{
a data.frame, 42 rows and 3 collumns (snp, chr, pos).
}
\usage{
data(ToyMap)
}
\description{
This data comprise 2 chromossomes with 2cM each and 21 
    monitored loci in each chromosome
}
