% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISQG.R
\docType{package}
\name{isqg-package}
\alias{isqg-package}
\alias{isqg}
\title{isqg: A package to perform \strong{in silico} quantitative genetics}
\description{
\code{isqg} provides R6/C++ classes for in silico quantitative 
    genetics. Mimic the meiosis recombination. Allows user defined extensions 
    which provides great flexibility. Details of the implementation are described 
    in \insertCite{g3;textual}{isqg}.
}
\references{
\insertAllCited{}
}
\author{
Fernando H. Toledo \email{f.toledo@cgiar.org}
}
\keyword{package}
