% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes.R
\docType{class}
\name{Specimen}
\alias{Specimen}
\title{Class providing object with methods to mimic in silico Specimens}
\value{
Objects of R6 class with methods to mimic in silico Specimens.
}
\description{
Mean to mimic a in silico Specimens. It is the working instances of
    the simulator.
}
\details{
Object of R6 class that points to C++ objetcs.
}
\section{Fields}{

\describe{
\item{\code{.ptr}}{External pointer to the instance of the C++ class Specimen.}
}}

