% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Classes.R
\docType{class}
\name{Specie}
\alias{Specie}
\title{Class providing object with methods to mimic in silico Genomes}
\value{
Objects of R6 class with methods to mimic in silico Genomes.
}
\description{
Mean to mimic a in silico Genomes. It is the machine instances of
    the simulator.
}
\details{
Object of R6 class that points to C++ objetcs.
}
\section{Fields}{

\describe{
\item{\code{.ptr}}{External pointer to the instance of the C++ class Specie.}
}}

